fix: terminar permisos de cliente normal
